requirements.txt
requests
feedparser
beautifulsoup4
lxml
python-dotenv
pyyaml
markdown
openai
playwright
pytest
